# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Group Chat Binder API",
    "description": "Modern group chat application built with Next.js, Prisma, and Supabase",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/auth/token": {
      "get": {
        "summary": "Get current user token",
        "description": "Retrieve the current user's authentication token from active session",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Token retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "No active session or authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Login and get token",
        "description": "Login with email and password to get authentication token",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, token returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Sign up new user",
        "description": "Create a new user account with email and password via Supabase",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing fields or user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/create-profile": {
      "post": {
        "summary": "Create user profile",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/friend-requests": {
      "get": {
        "summary": "Get friend requests",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "received"
              ]
            },
            "description": "Type of friend requests to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "List of friend requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendRequest"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Send friend request",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendRequestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Friend request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token"
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token"
          },
          "token_type": {
            "type": "string",
            "description": "Token type (usually Bearer)"
          },
          "expires_at": {
            "type": "integer",
            "description": "Token expiration timestamp"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration time in seconds"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "User ID"
              },
              "email": {
                "type": "string",
                "description": "User email"
              },
              "username": {
                "type": "string",
                "description": "Username"
              }
            }
          },
          "instructions": {
            "type": "object",
            "properties": {
              "usage": {
                "type": "string",
                "description": "How to use the token"
              },
              "example": {
                "type": "string",
                "description": "Authorization header example"
              },
              "expires": {
                "type": "string",
                "format": "date-time",
                "description": "Token expiration date"
              }
            }
          }
        },
        "required": [
          "access_token",
          "token_type",
          "user"
        ]
      },
      "SignupRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password (min 6 characters)"
          },
          "username": {
            "type": "string",
            "description": "Optional username (defaults to email prefix)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "access_token": {
            "type": "string",
            "description": "JWT access token (only if email confirmation disabled)"
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token (only if email confirmation disabled)"
          },
          "token_type": {
            "type": "string",
            "description": "Token type (usually Bearer)"
          },
          "expires_at": {
            "type": "integer",
            "description": "Token expiration timestamp"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token expiration time in seconds"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "User ID"
              },
              "email": {
                "type": "string",
                "description": "User email"
              },
              "username": {
                "type": "string",
                "description": "Username"
              }
            }
          },
          "confirmation_required": {
            "type": "boolean",
            "description": "Whether email confirmation is required"
          },
          "instructions": {
            "type": "object",
            "properties": {
              "usage": {
                "type": "string",
                "description": "How to use the token (if provided)"
              },
              "example": {
                "type": "string",
                "description": "Authorization header example (if token provided)"
              },
              "expires": {
                "type": "string",
                "format": "date-time",
                "description": "Token expiration date (if token provided)"
              },
              "next_step": {
                "type": "string",
                "description": "Next step for confirmation (if required)"
              },
              "login_after": {
                "type": "string",
                "description": "How to login after confirmation (if required)"
              }
            }
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ]
      },
...
      "FriendRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fromUserId": {
            "type": "string"
          },
          "toUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "DECLINED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fromUser": {
            "$ref": "#/components/schemas/User"
          },
          "toUser": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "id",
          "fromUserId",
          "toUserId",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateProfileRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username"
        ]
      },
      "CreateProfileResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
...
      "SendFriendRequestRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "anyOf": [
          {
            "required": [
              "email"
            ]
          },
          {
            "required": [
              "username"
            ]
          }
        ]
      },
...
    }
  }
}